analisi codice

+++ src/main/java/org/generation/italy/hairhub/model/entities/SalonTreatment.java
@@ -22,11 +22,11 @@ public class SalonTreatment {
     @JoinColumn(name = "serv_id", nullable = false)
     private Treatment treatment;
 
-    private BigDecimal price;
+    private double price;
 
     public SalonTreatment(){}
 
-    public SalonTreatment(Long id,Salon salon, Treatment treatment, BigDecimal price) {
+    public SalonTreatment(Long id,Salon salon, Treatment treatment, double price) {
         this.id = id;
         this.salon = salon;
         this.treatment = treatment;
@@ -42,7 +42,7 @@ public class SalonTreatment {
     public Treatment getTreatment() {
         return treatment;
     }
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -52,7 +52,7 @@ public class SalonTreatment {
     public void setTreatment(Treatment treatment) {
         this.treatment = treatment;
     }
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
     public void setId(long id) {
